test: Add integration test cases
