+++ package.json
@@ -63,10 +63,10 @@
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-storybook": "^0.6.13",
     "husky": "^8.0.3",
-    "jest": "^29.6.2",
+    "jest": "^29.6.4",
     "jest-dom": "^4.0.0",
     "jest-environment-jsdom": "^29.6.2",
-    "msw": "^1.2.1",
+    "msw": "^1.3.0",
     "prettier": "^3.0.2",
     "storybook": "^7.4.0-alpha.0",
     "ts-jest": "^29.1.1"
