+++ src/api/client.ts
@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-explicit-any */
 /* eslint-disable no-unused-vars */
 import axios from 'axios';
 import { getStorageItem } from '@/utils/storage';
