+++ src/app/posts/[post_id]/page.tsx
@@ -1,7 +1,6 @@
 'use client';
 
 import Link from 'next/link';
-import { styled } from 'styled-components';
 import { useEffect, useState } from 'react';
 import { Renderer } from '@/components/markdown/Renderer';
 import { postAPI } from '@/api/post';
@@ -9,43 +8,8 @@ import { Post } from '@/types/Post';
 import { beforeDateFormat } from '@/utils/dateFormat';
 import { GiscusComment } from '@/components/GiscusComment';
 import { LoadingSkeleton } from '@/components/loading/LoadingSkeleton';
-import { CategoryCard } from '@/components/card/CategoryCard';
-
-const S = {
-  container: styled.div`
-    ${({ theme }) =>
-      theme.MIXINS.flexBox('column', 'flex-start', 'flex-start')};
-    height: 100%;
-    width: 100%;
-    padding: 20px; 16px;
-  `,
-  heading: styled.h1`
-    font-weight: bold;
-    font-size: 32px;
-    color: ${({ theme }) => theme.colors.text['222222']};
-  `,
-  datetime: styled.p`
-    font-size: 14px;
-    width: 100%;
-    text-align: right;
-    color: ${({ theme }) => theme.colors.gray['747474']};
-    margin-top: 12px;
-  `,
-  section: styled.section`
-    width: 100%;
-    margin-top: 48px;
-  `,
-  tags: styled.ul`
-    border-top: 1px solid #eaecef;
-    padding-top: 16px;
-    flex-wrap: wrap;
-    ${({ theme }) => theme.MIXINS.flexBox('row', 'center', 'flex-start')};
-
-    li {
-      margin-top: 12px;
-    }
-  `,
-};
+import { TagCard } from '@/components/card/TagCard';
+import * as S from './page.style';
 
 export default function PostPage({ params }: { params?: { post_id: string } }) {
   const [postData, setPostData] = useState<Post>();
@@ -84,11 +48,11 @@ export default function PostPage({ params }: { params?: { post_id: string } }) {
       <S.section>
         <Renderer data={postData.content} />
         <S.tags>
-          {postData.tags.split('#').map(tag => {
+          {postData.tags.map(tag => {
             if (!tag) return null;
             return (
-              <Link key={tag} href={`/categories/${tag}`}>
-                <CategoryCard categoryName={`#${tag}`} />
+              <Link key={tag.name} href={`/tags/${tag.name}`}>
+                <TagCard tagName={tag.name} />
               </Link>
             );
           })}
