+++ src/types/Tag.ts
@@ -1,4 +1,4 @@
-export interface Category {
+export interface Tag {
   id: number;
   name: string;
 }
